// frontend/ai-news-ui/src/shared/interfaces/analysis/ExecutiveOverview/TopicRadarType.ts
interface  TopicDistribution{
  name: string;
  count: number;
}

export interface TopicTooltipItem {
  name: string;
  value: string | number;
}
export interface TopicRadarType {
  count: number;
  distribution: TopicDistribution[];
}

export interface TopicRadarLegendTooltip {
  title: string;
  items: TopicTooltipItem[];
}